# 06_API/OpenAPI/Chalo_Farva_API_v1.0.yaml
openapi: 3.0.3
info:
  title: Chalo Farva Master API Engine
  description: Complete production OpenAPI specification for Chalo Farva - Gujarat-first AI-powered travel platform.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
  - url: https://api.chalofarva.com/v1
    description: Production API Server
paths:
...
  /destinations:
    get:
      summary: Discover & Filter Gujarat Destinations
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of active Gujarat destinations
  /destinations/{slug}:
    get:
      summary: Get Destination Details by Slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Destination detailed record
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT